Add FeatureTable component for URL submission and integrate it into the popup
